% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitStats.R
\name{FitStats}
\alias{FitStats}
\title{Calculate Fit Statistics}
\usage{
FitStats(y, P, r = NULL, statistics = c("rmse", "mad", "mape", "meape",
  "rmsle", "mrae"))
}
\arguments{
\item{y}{A vector with the observable values in the dataset.}

\item{P}{A matrix with the the predicted values.}

\item{r}{A vector with naive predictions (optional).}

\item{statistics}{A list of statistics to be calculated.}
}
\value{
A matrix with the selected statistics.
}
\description{
Calculate Fit Statistics
}
\details{
The user can specify the following statistics: RMSE, MAD, MAPE, MEAPE, RMSLE, and MRAE.
 If "mrae" is specified by the user, a vector r must be provided. Otherwise, MRAE. will be not calculated and
 a warning messenger will be shown (The remaining statistics will be calculated in any case).
 If P or y contain negative numbers, RMSLE will be not calculated and a warning messenger will be 
 shown.
}
\examples{
x <- rpois(100, lambda = 100)
Y <- 4 + 3 * x + rnorm(100)
P <- matrix(predict(lm(Y ~ x)), ncol = 1)
r <- sample(min(Y):max(Y), 100, replace = TRUE)
FitStats(y = Y, P = P, r = r)
}
\author{
Patrick C. Silva
}

